[Android] Introduce sharedContext for ARCore camera preview rendering & integrate it with THREE.js

+++ packages/expo-ar/android/src/main/res/raw/camera_vertex_shader.vert
@@ -1,6 +1,6 @@
-#version 300 es
+#version 320 es
 
-precision highp float;
+precision mediump float;
 
 in vec2 aPosition;
 in vec2 aTextureCoord;
